+++ tests/Feature/RoutingTest.php
@@ -28,16 +28,15 @@ it('generates localized routes', function () {
 it('generates localized routes with {lang} prefix', function () {
     app(Translator::class)->addPath(__DIR__ . '/../Support/lang');
 
-    Route::group([
-        'prefix' => '{lang}',
-        'lang' => ['fr', 'de', 'es', 'it'],
-    ], function () {
-        Route::get('/example', fn () => 'Hello, World!')
-            ->lang([
-                'it' => '/esempio',
-                'es' => '/ejemplo',
-            ]);
-    });
+    Route::lang(['fr', 'de', 'es', 'it'])
+        ->prefix('{lang}')
+        ->group(function () {
+            Route::get('/example', fn () => 'Hello, World!')
+                ->lang([
+                    'it' => '/esempio',
+                    'es' => '/ejemplo',
+                ]);
+        });
 
     app(L10n::class)->registerLocalizedRoutes();